A Swift XML parser (uses Cocoa APIs)
This is a working example of an XML parser written in Swift but using
Objective-C APIs to handle the heavy lifting. It’s based partly on one
of Apple’s own example applications. It is presently designed to read a
specific XML file but could be generalized and made much more Swifty.
It’s only a preliminary experiment from someone who isn’t a developer
and hasn’t done much coding lately.